requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
multidict==6.6.3
openai==1.98.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
readability-lxml==0.8.4.1
regex==2025.7.34
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
